+++ src/lib.rs
@@ -402,6 +402,8 @@ macro_rules! bitfield {
 #[doc(hidden)]
 pub use core::convert::Into;
 #[doc(hidden)]
+pub use core::convert::TryInto;
+#[doc(hidden)]
 pub use core::fmt;
 #[doc(hidden)]
 pub use core::mem::size_of;